Input C# Script
A new script was created to manage the horizontal inputs received when playing the game. These inputs are then implemented through code to make the Spaceship player to move left and right horizontally with the corresponding horizontal input received. This script is modular allowing for any object to correctly move horizontally with input as long as the Movement script and SpeedDirectionController (previously named SpeedController) script are allied to that same object. This script references the Movement script. This script was also included in the inspector for the Spaceship.