Enemy and Bullet Prefab Editing Values and Adding Movement Scripts
The Movement script and SppedDirectionController script was added to the enemy prefab in unity inspector. The enemy Acceleration variable in the Movement script was set to 150. The Direction in the SpeedDirectionController script was set to x=0 and y=-1.

The Movement script and SpeedDirectionController script was also added to the Bullet prefab in the unity inspector. The Bullet Acceleration variable in the Movement script was set to 200. The Direction in the SpeedDirectionController script was set to x=0 and y=1.

The Acceleration of the Spaceship player in the Movement script was set to 5000 and the Movement in the SpeedDirectionController was set to the Spaceship movement.