Shooting C# Script Improvement
The input for shooting the bullet was in the ShootingScript script and since an InputManager script exists the input required for shooting was configured into the InputManager. The InputManager now accounts for input of the left mouse button and the space bar to act out a shoot function within the ShottingScript script. In the ShootingScript script a function called Shoot was created to shoot the bullet out of the Spaceship player when the Fire1 (left mouse) or Jump (space bar) has been triggered. The scripts so far are working as intended.