Velocity Implementation
Velocity was added into the Movement c# script based on the BulletMoveForward script. Therefore the BulletMoveForward script was no longer required and therefor it was deleted. The values of Initial Velocity and Acceleration were continuously changed and tweaked in the unity inspector of the Enemy, Bullet, and Spaceship so that the game runs smoothly.