Velocity and Commenting
Velocity in the Movement C# script was improved where the direction is a changeable variable rather than having positive and negative velocity values there will be positive and negative Y values to determine the direction of the velocity vector. The C# scripts have been effective commented and the variables have been named correctly.