Velocity Fixed and Double Checking
Double checked that everything was running smoothly before submission. A few values were changed in the Enemy and Spaceship prefabs inspector in unity regarding Velocity and Acceleration in the Movement script. Overall, the game is running as intended with velocity and acceleration working for the Enemy and Bullets. There is no velocity on the Spaceship, however the values to change velocity in the Spaceship inspector under the Movement and SpeedDirectionController script are still present.