Pickup Spawning and Script
The Pickup prefab will spawn overtime using the SpawnOverTimeScript applied to the Spawner in the Scene. The pickup object is also destroyed when it hits the player, although this may cause complications later in the Pickup script. The DestroyedOnCollision script was also modified in the inspector with the Bullet and the EnemySpaceship to change the whitelist tagged objects and the blacklist tagged objects.