MoveConstantly script
The MoveConstantly script is the same as the Movement script and the SpeedDirectionController script but in one script. Therefore the MoveConstantly script was added to the Bullet and the Enemy with the values slightly modified to make the game easier. Therefore the Movement script and the SpeedDirectionController script were deleted and replaced. Also the LeftBoarder and the RightBorder in the scene were increased in size on the y-axis to cover the whole sides of the scene so that the EnemySpaceships do not fly out of bounds.