UI Management and Blacklisting
The UI scene was loaded where the player has a health bar that goes down when the player gets hit by enemy bullets. The blacklisting required modification due to some complications in the code like the player disappearing when hit by an enemy spaceship when I want damage to be done instead, and then the player would just hold and stop the spaceship. Therefore the DestoryedOnCollision Script was added to the Player Spaceship, where the Enemy and Pickup tags are blacklisted. There is still quite a bit of work required with the UI and the Machine gun weapon needs serious attention as it is messing with the bullets.