Removed ShootingScript and Updated the InputManager Script
It was apparent that the ShootingScript and the WeaponMachineGun script were very similar and that they were performing the same function. This caused the error with the bullets because both scripts were effectively being used at the same time. In the ShootingScript there was one reference of the Shoot method in the InputManager script, therefore I changed the reference to refer to the WeaponBase Shoot method instead. Therefore the ShootingScript was deleted and the WeaponBase Shoot method replaced the ShootingScript because it contains both MachineGun and TripleShot Shoot methods.