DamageOnCollision was added to all Enemies and an EnemyBoss Prefab was Created
I felt like the EnemySpaceship and the Rock (previously named enemy) should deal damage to the player on collision rather than the object just being destroyed. Therefore the DamageOnCollision script was added to the corresponding prefabs in Unity and varying amounts of damage inflicted on the player were put into the serialized field. The EnemyBoss prefab is just an enlarged EnemySpaceship that moves slower due to the limited sprite assets provided. This EnemyBoss will have the feature to use a triple shot.