Experimenting with Enemy Health and Enemy Bullets
Various methods were trialed to try to get the Enemy Health working. A canvas was created and attached to the Enemy and Boss spaceship prefabs that holds a red health bar. Unfortunately this bar does not seem to decrease when the enemy is hit with a player bullet, they are instantly destroyed on contact. I might need to change the collision detection somehow. A new IDamage and IHeal interfaces were created to be used instead of the IHealth because when applying the IHealth to the EnemyHealth script healing is required within the script, however, the enemies will not hold the capabilities to heal. Although, upon reading further into the task it is desirable for the EnemyHealth to use the IHealth Interface. Therefore they will be removed in the next commit, but this is all the time I have for now.