Redundant InputManager Script Removed
After further inspection the InputManager script is the same as the PlayerInput script minus the weapon change. Therefore the InputManager Script was removed from the player and deleted as the script is not required because the PlayerInput script   does the exact same tasks that the InputManager script completes. However, the PlayerInput has the addition of being able to change the weapons.