Enemy Boss Triple Shot Working
The enemy boss has a triple shot that when shot the bullets travel downwards towards the player rather than upwards, like they were doing previously. To achieve this, in the WeaponTripleShot script an if statement was created where if the tag is "Player" then the triple shots will be fired upwards due to the y value in the 2D vector being a positive 0.5 float. then an else statement was created meaning that any other object using the triple shot function that is not tagged as "Player" will shoot downwards, due to the y value in the 2D vector being a negative 0.5 float. Therefore all of the shooting for both the player and all enemy spaceships (including the Boss) are working perfectly.