Pickup Timer/Cooldown Added
When the player picks up the triple shot power up they shoot the triple shot weapon constantly, the weapon never reverts back to the machine gun. Therefore a timer/cooldown was added for the powerup/pickup in the PlayerInput script. The PlayerInput script deals with the SwapWeapon method and therefore it is perfect to add a timer here that directly involves the player as well. A serialized field was added for a private powerUpDuration float that is then editable in unity under the player depending on how long the cooldown is desired. Another variable was made to be able to reset the timer. Under void Update an if statement was created that when the powerUpTimer is greater than 0 then it will start to count down using time of seconds. another if statement is made so that when the timer has run out the weapon will swap back to the machine gun. In the SwapWeapon method, when the weapon is swapped to the triple shot the timer is rest by setting the powerUpTimer to the powerUpDuration that is ready to start counting down on the next frame update. This works perfectly giving the player 5 seconds to use the triple shot powerup after picking up the pickup before reverting back to the normal weapon.