Player Healing Achieved and Healing Pickup Added
The player now receives health when it picks up the HealPickup. This pickup was coloured green and spawns over time in the spawner by attaching the SpawnOverTime script linked to the prefab and set a time. In the PlayerHealth script under the Heal method, an adjustment was required where the UpdatePlayerHealthSlider method was called from the UIManager to update the player health bar at the top of the screen when playing to show the health increase. in the Pickup script a Serialized field was created of the healAmount at which the HealPickup will heal the player by, therefore the amount can be changed per request. Under the HandlePlayerPickup method and else if statement was created where if the pickup has the tag "Heal" then the Heal method from IHealth will be called.