Enemy Health Bar and Damage Running Smoothly
The enemy health bar is working perfectly as intended. The health bar is very similar to the Player health bar in PlayerHealth. The only difference is that in EnemyHealth there is a serialize field that is liked to the slider in the emeny and boss prefab that is updated when damage is taken rather than having to go into the UIManager. The DestroyedOnCollision script was experimented with in the enemy spaceship prefabs to determine what tags should be blacklisted and whitelisted. In conclusion, the DestroyedOnCollision script was removed from the enemy and boss spaceships and this allowed the bullet to cause damage to the spaceships rather than just destroying them on collision. However, the boulder was kept as destroyed on collision by the bullet because the boulder does not take damage.