Coding Standards Applied and Final Touch Ups Made
The coding standards were applied by commenting thoroughly were required and double checking that all parenthesis are correct and variables are named appropriately and accordingly. Final touches such as making the speed of the enemies and player desirable, adjusting the fire delay, and the spawn time of enemies were experimented and then applied. I final run through of the game was also conducted and everything is running smoothly and as intended.